% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModel.R
\name{linearModel}
\alias{linearModel}
\title{linearModel}
\usage{
linearModel(X, Y)
}
\arguments{
\item{X}{Design matrix with covariate values in it.}

\item{Y}{Observed Y, each row corresponds to one outcome.}
}
\value{
A dataframe with estimates of coefficients, standard errors, t-values and p-values.
}
\description{
Provided the same information that summary (lm) does in base R
This package does not have the functionality to include interactions,
higher order terms, or functions of variables. Those must be included as
part of the design matrix.
Furthermore, all data that is put into the function is assumed to be a
part of the regression. There is no variable selection process.
}
\examples{
set.seed(1616)
linearModel(X = matrix(rnorm(500), ncol = 5, nrow = 100), Y = rnorm(100))

}
